feat: when testing, add DB_FILTER=devel
Multidb environments are harder to test because any http test lands in the db selector instead of the actual instance.

The solution is to add a dbfilter, but that would become problematic in devel when doing db backups or resetdb.

Thus, when `invoke test`ing, add automatically the DB_FILTER, so there's one less thing to worry about.